Move interval to domain
